test: move more utils to utils.ts
